cohorts/2025/04-analytics-engineering/github-to-bq-pipeline/requirements.txt
apache-airflow-providers-google==10.15.0
google-cloud-storage==2.13.0
google-cloud-bigquery==3.19.0
requests==2.31.0